+++ Makefile
@@ -1,0 +1,34 @@
+CC = g++
+CFLAGS = -std=c++11 -Wall
+TARGET = program
+
+BUILD_DIR = ./build
+SRC_DIR = ./src
+INCLUDE_DIR = ./include
+
+run: ${TARGET}
+	${TARGET}
+
+${TARGET}: ${BUILD_DIR}/Apartamento.o ${BUILD_DIR}/Casa.o ${BUILD_DIR}/Cobertura.o ${BUILD_DIR}/Imovel.o ${BUILD_DIR}/Cliente.o ${BUILD_DIR}/main.o
+	${CC} ${CFLAGS} ${BUILD_DIR}/*.o -o ${TARGET} 
+
+${BUILD_DIR}/Imovel.o: ${INCLUDE_DIR}/Imovel.hpp ${SRC_DIR}/Imovel/Imovel.cpp
+	${CC} ${CFLAGS} -I ${INCLUDE_DIR}/ -c ${SRC_DIR}/Imovel/Imovel.cpp -o ${BUILD_DIR}/Imovel.o
+
+${BUILD_DIR}/Apartamento.o: ${INCLUDE_DIR}/Apartamento.hpp ${SRC_DIR}/Imovel/Apartamento.cpp
+	${CC} ${CFLAGS} -I ${INCLUDE_DIR}/ -c ${SRC_DIR}/Imovel/Apartamento.cpp -o ${BUILD_DIR}/Apartamento.o
+
+${BUILD_DIR}/Casa.o: ${INCLUDE_DIR}/Casa.hpp ${SRC_DIR}/Imovel/Casa.cpp
+	${CC} ${CFLAGS} -I ${INCLUDE_DIR}/ -c ${SRC_DIR}/Imovel/Casa.cpp -o ${BUILD_DIR}/Casa.o
+
+${BUILD_DIR}/Cobertura.o: ${INCLUDE_DIR}/Cobertura.hpp ${SRC_DIR}/Imovel/Cobertura.cpp
+	${CC} ${CFLAGS} -I ${INCLUDE_DIR}/ -c ${SRC_DIR}/Imovel/Cobertura.cpp -o ${BUILD_DIR}/Cobertura.o
+
+${BUILD_DIR}/Cliente.o: ${INCLUDE_DIR}/Cliente.hpp ${SRC_DIR}/Cliente/Cliente.cpp
+	${CC} ${CFLAGS} -I ${INCLUDE_DIR}/ -c ${SRC_DIR}/Cliente/Cliente.cpp -o ${BUILD_DIR}/Cliente.o
+
+${BUILD_DIR}/main.o: ${INCLUDE_DIR}/*.hpp ${SRC_DIR}/main.cpp
+	${CC} ${CFLAGS} -I ${INCLUDE_DIR}/ -c ${SRC_DIR}/main.cpp -o ${BUILD_DIR}/main.o
+
+clean:
+	rm -f ${BUILD_DIR}/*
