+++ src/Cliente/Cliente.cpp
@@ -2,13 +2,11 @@
 #include <string>
 #include "Cliente.hpp"
 
-void Cliente::print(){
-
+void Cliente::imprimeDados(){
 	std::cout << "  Nome: " << nome << endl
 		 << "  Telefone: " << telefone << endl
 		 << "  Endereço: " << endereco << endl
 		 << "  Cidade: " << cidade << endl
 		 << "  Estado: " << uf << endl
 		 << "  CEP: " << cep << endl;
-
 }
