+++ src/Imovel/Apartamento.cpp
@@ -1,21 +1,21 @@
-#include "Cobertura.hpp"
+#include "Apartamento.hpp"
 
-double Cobertura::valor() {
+double Apartamento::valor() {
     return area * valorMetroQuadrado;
 }
 
-double Cobertura::comissao(){
-    return area * valorMetroQuadrado * COMISSAOCOBERTURA;
+double Apartamento::comissao(){
+    return area * valorMetroQuadrado * COMISSAOAPARTAMENTO;
 }
 
-void Cobertura::print(){
-    std::cout << "[Cobertura]" << endl;
-    Imovel::print();
+void Apartamento::imprimeApartamento(){
+    std::cout << "[Apartamento]" << endl;
+    Imovel::imprimeImovel();
     std::cout << "Area: " << area << endl
             << "  Quartos: " << numQuartos << endl
             << "  Banheiros: " << numBanheiros << endl
             << "  Vagas: " << numVagas << endl
-            << "Taxa de Comissão: " << 10 << "%" << endl
+            << "Taxa de Comissão: " << 4 << "%" << endl
             << "Valor Comissão: R$ " << fixed << setprecision(2) << valorComissao << endl
             << "Valor de Venda: R$ " << fixed << setprecision(2) << valorImovel << endl;
 }
